Added Payment and order Service (in-complete)

+++ product-service/src/main/kotlin/com/ethyllium/productservice/dto/request/CreateProductRequest.kt
@@ -6,7 +6,7 @@ import java.math.BigDecimal
 data class CreateProductRequest(
     val name: String,
     val description: String,
-    val price: BigDecimal,
+    val price: Double,
     val discount: BigDecimal? = null,
     val sellerId: String,
     val attributes: Map<String, String> = mapOf(),
